1단계 - List<Section> sections를 일급 컬렉션으로 변경
